Util/nao2gto/requirements.txt
numpy>=1.11.0
matplotlib>=1.5.1
ruamel.yaml>=0.15.34
scipy>=0.17.0
